[THUNDER-48] Explicitly specify Hadoop dependency
While the build instructions for the scala code are not yet explicit, sbt does a bit of magic to pull the Hadoop version from Spark, which defaults to 1.0.4 at the moment.  This change at least makes the version explicit, so the user can know to change it if desired.

Fixes #48 (kinda)