change keys for images with nplanes to integer index
Initially with nplanes specified, keys for image rdds were (fileidx,
pointInFile) integer pairs. However this causes problems downstream
in blocks conversion. This commit changes that back to straight integer
indices. The downside is that for tifs, the entire tif must be loaded
into memory in order to count the number of planes, whereas previously
this could be pipelined lazily.