first attempt at a local aggregate by regions method
This version is on Series objects, and takes as an argument a
sequence of sequences of series keys, with each sequence of keys
specifying a region. Each region is put onto its own worker
through groupByKey, then the series values are accumulated
into a local 2D array and the passed aggregation function applied
to the array.