add Images meanByRegions
Adds a meanByRegions method on Images, roughly analogous to the
existing method on Series objects. This version returns a new
Images object with shape (1, nRegions). Each of the nRegions
values is the result of taking the mean within the original image
array over all pixels with a corresponding nonzero mask value.