change default thunder version in ec2 script
Previous default was to install Thunder from the current head of
the master branch on github. Now this will still be the default
if the local install of Thunder is a development version ('_dev'
in version string); otherwise the default will be to install on ec2
the same released version as is installed locally.