Simple support for conf files
Rather than provide the ability to load all image loading configuration
parameters from a file, this simply adds support for loading the
primary attributes for which flat binary images cannot be read without,
namely dtype and dimensions. These are files that are useful to keep
with the data (akin to a header), whereas other parameters like
recursive, ext, etc can be inferred or are environment specific.