Set index attribute explicitly, bypassing setter
Avoids calling the index setter, which forces a first call, in the case
that the index is being set explicitly. This does skip the check on
index length, but we assume that if the index is being set explicitly
during Series construction, the developer or user is responsible for
using something with the correct length.